Move some types to common
