+++ sdk/types/result.go
@@ -423,24 +423,6 @@ type SelectSubjectById struct {
 
 type IncludeAll = map[string]interface{}
 
-type Link struct {
-	Href             string `json:"href" yaml:"href"`
-	MediaType        string `json:"media-type,omitempty" yaml:"media-type,omitempty"`
-	Rel              string `json:"rel,omitempty" yaml:"rel,omitempty"`
-	ResourceFragment string `json:"resource-fragment,omitempty" yaml:"resource-fragment,omitempty"`
-	Text             string `json:"text,omitempty" yaml:"text,omitempty"`
-}
-
-type Property struct {
-	Class   string `json:"class,omitempty" yaml:"class,omitempty"`
-	Group   string `json:"group,omitempty" yaml:"group,omitempty"`
-	Name    string `json:"name" yaml:"name"`
-	Ns      string `json:"ns,omitempty" yaml:"ns,omitempty"`
-	Remarks string `json:"remarks,omitempty" yaml:"remarks,omitempty"`
-	UUID    string `json:"uuid,omitempty" yaml:"uuid,omitempty"`
-	Value   string `json:"value" yaml:"value"`
-}
-
 type Resource struct {
 	Base64      *Base64         `json:"base64,omitempty" yaml:"base64,omitempty"`
 	Citation    *Citation       `json:"citation,omitempty" yaml:"citation,omitempty"`
